Make ParsedLine inherit from Queue<string> Move to ParsedLine and ParsedFile methods previously located in FileReader.
